[FLINK-8302][table]Minor update to clean code of SHIFT_LEFT and SHIFT_RIGHT functions

+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/functions/sql/ScalarSqlFunctions.scala
@@ -63,7 +63,7 @@ object ScalarSqlFunctions {
     SqlKind.OTHER_FUNCTION,
     ReturnTypes.ARG0,
     null,
-    OperandTypes.family(SqlTypeFamily.NUMERIC, SqlTypeFamily.NUMERIC),
+    OperandTypes.family(SqlTypeFamily.NUMERIC, SqlTypeFamily.INTEGER),
     SqlFunctionCategory.NUMERIC
   )
 
@@ -72,7 +72,7 @@ object ScalarSqlFunctions {
     SqlKind.OTHER_FUNCTION,
     ReturnTypes.ARG0,
     null,
-    OperandTypes.family(SqlTypeFamily.NUMERIC, SqlTypeFamily.NUMERIC),
+    OperandTypes.family(SqlTypeFamily.NUMERIC, SqlTypeFamily.INTEGER),
     SqlFunctionCategory.NUMERIC
   )
